+++ Messages.js
@@ -1,0 +1,8 @@
+const messages = [
+  {
+  id: 0,
+  from: "Bart",
+  text: "Welcome to CYF chat system!",
+  },
+];
+module.exports = messages;
